Fix tests Breadcrumbs are specific to releases
